+++ aws/rs_aws_vpc/changelog.md
@@ -1,4 +1,10 @@
 VPC Plugin changelog
+v1.2
+- adding action enablevpcclassiclink, disablevpcclassiclink to vpc resource
+- adding action enablevpcclassiclinkdnssupport,disablevpcclassiclinkdnssupport to vpc resource
+- adding action create_tag, delete_tag to vpc resource
+- adding show action to vpc resource
+- adding tags resource to vpc plugin
 
 v1.1
 -----